relocate script
