+++ src/webmaster/modules/WMTableManager.py
