--- src/webmaster/parsers/unactive_domens.txt
@@ -1,67 +1,0 @@
-zima.narkology24-7.ru
-gusev.narkology24-7.ru
-semikarakorsk.narkology24-7.ru
-uren.narkology24-7.ru
-zvenigovo.narkology24-7.ru
-aleksandrovsk-sahalinskiy.narkology24-7.ru
-mihaylov.narkology24-7.ru
-udachnyy.narkology24-7.ru
-zakamensk.narkology24-7.ru
-plievo.narkology24-7.ru
-surhahi.narkology24-7.ru
-zelenchukskaya.narkology24-7.ru
-murmashi.narkology24-7.ru
-kurya.narkology24-7.ru
-krutiha.narkology24-7.ru
-ust-charyshskaya-pristan.narkology24-7.ru
-ivanovskaya.narkology24-7.ru
-nizhnebakanskaya.narkology24-7.ru
-kurskaya.narkology24-7.ru
-balakirevo.narkology24-7.ru
-doneck.narkology24-7.ru
-gorlovka.narkology24-7.ru
-debalcevo.narkology24-7.ru
-dokuchaevsk.narkology24-7.ru
-enakievo.narkology24-7.ru
-zhdanovka.narkology24-7.ru
-kirovskoe.narkology24-7.ru
-makeevka.narkology24-7.ru
-snezhnoe.narkology24-7.ru
-torez.narkology24-7.ru
-harcyzsk.narkology24-7.ru
-shahtyorsk.narkology24-7.ru
-yasinovataya.narkology24-7.ru
-lugansk.narkology24-7.ru
-alchevsk.narkology24-7.ru
-antracit.narkology24-7.ru
-bryanka.narkology24-7.ru
-kirovsk-luganskayaobl.narkology24-7.ru
-krasnodon.narkology24-7.ru
-krasnyj-luch.narkology24-7.ru
-pervomajsk.narkology24-7.ru
-rovenki-luganskayaobl.narkology24-7.ru
-sverdlovsk.narkology24-7.ru
-stahanov.narkology24-7.ru
-alyoshki.narkology24-7.ru
-berislav.narkology24-7.ru
-genichesk.narkology24-7.ru
-golaya-pristan.narkology24-7.ru
-kahovka.narkology24-7.ru
-novaya-kahovka.narkology24-7.ru
-skadovsk.narkology24-7.ru
-tavrijsk.narkology24-7.ru
-herson.narkology24-7.ru
-berdyansk.narkology24-7.ru
-vasilevka.narkology24-7.ru
-volnyansk.narkology24-7.ru
-gulyajpole.narkology24-7.ru
-dneprorudnoe.narkology24-7.ru
-zaporozhe.narkology24-7.ru
-kamenka-dnepr.narkology24-7.ru
-melitopol.narkology24-7.ru
-molochansk.narkology24-7.ru
-orekhov.narkology24-7.ru
-pologi.narkology24-7.ru
-primorsk.narkology24-7.ru
-tokmak.narkology24-7.ru
-energodar.narkology24-7.ru
